Reduce Authorized client API
