+++ core/src/test/java/com/jaspersoft/android/sdk/network/SpringAuthServiceTest.java
@@ -53,7 +53,7 @@ import static org.mockito.Mockito.*;
 })
 public class SpringAuthServiceTest {
     @Mock
-    AuthenticationRestApi mRestApi;
+    AuthRestApi mRestApi;
     @Mock
     JSEncryptionAlgorithm mAlgorithm;
     @Mock
